Created _lines and border methods
The first step towards being able to create borders around a table. The
border method will create a border line when passed in a JointChars
instance and a list of lines (which can be created using the _lines
method). Effectively this means that top and bottom borders can be
created now, but still no side borders.