Removed unecessary iter() call in div_lines()
No reason to make something a list, then call iter, when you can just
zip and iter through it. I'll probably change it back to make the
h_chars variable an iterator instead of a list at some point because I
don't know that there's a reason to make it a list since it's just used
once and then disposed of.